pkg/kubectl: check version of kubectl
If someone tries to execute krew binary standalone with kubectl <1.12, adding
a failure.

I'm not entirely sure if it's a valid use case. The only encounter of this
could be when someone installs krew, and directly tries to invoke the binary
(as krew 0.2 won't work on kubectl 1.11 or older).

Probably harmless for now (also added a KREW_DISABLE_KUBECTL_VERSION_CHECK
escape hatch for dev/build purposes). We can remove this pkg and check soon.
Currently adds ~100 ms to each krew cmd.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>