Allow mocking of url downloads with local file
- rename --source to --manifest in install command.
- introduce --archive flag to install command to override the installation from
  the uri specified in the manifest file.
- remove getFileFromArg() method (not needed)

This patch allows specifying both plugin yamls and archives to install cmd:

    krew install --manifest foo.yaml --archive foo.zip

It does not download a file, but it still checks integrity (so it can be used to
check the manifest without uploading the archive anywhere).

The method signatures got very long but I think that's partially because
we probably didn't design the methods in install.go well (the code smell here
is the repetitive args passed from one method to another all the way down).

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>