pkg/index: add manifest apiVersion check (#72)
We currently do not check apiVersion in plugin manifests and fail other
validation steps like "bin field is missing". This is confusing to users.

Adding an apiVersion check. Right now it's an exact match, but it should be ok.
When we can figure out how to parse k8s apiVersion format and compare them,
we can refactor this into isMinAPIversion().

This will require updates to many testdata files next time apiVersion is bumped.