Add basic integration tests (#203)
* Add util to set up krew integration tests

* Add utilities to initialize the krew index

For faster tests, the index is only cloned once per test run.
Additionally, the index is peristently cached between test runs.

* Add simple integration tests to cover basic krew commands

* Review comments

* Remove flawed test cases

* Add basic integration test for `krew help`

* Setup PATH variable in tests so that installed plugins can be executed

Also, optionally link to a krew binary as specified via the KREW_BINARY env varible.

* Make it possible to call plugins other than `krew`

* Enable integration tests on travis

* Fix boilerplate header

* Move IndexURI to constants

This is important, because importing the URI in integration tests
triggers the init functions in the cmd package. This has very nasty
and hard to find side-effects.

* Review comments

- rename package integration -> test
- disable integration tests in run-tests.sh
- remove obsolete tempdir helper

* Enable integration tests on travis

- Use a script run-integration-tests.sh for that
- Add instructions how to run this script locally

* Rename test types to avoid stuttering