hack: verify builds are installable
This script tests if a build is installable (as documented in the readme)
from its tarball+yaml file.

I felt like adding this as extra confidence on all the refactoring we're doing,
and I know we have integration tests that cover this. Since installation is
almost like our API, I thought it would be good to test.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>