test: remove -v from tests (#285)
-v is only useful if you want to see all the test cases that are being run.
I don't think this is useful locally or in the build system. I've been having
trouble scrolling a lot and trying to find the FAIL: line.

Any failed test will still print its t.Log or stdout/stderr printing along with
the test method name, so this should be fine.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>