Clean up stale krew installations on windows (#314)
* Clean up stale krew installations on windows

This introduces a preRun check to root cmd. On windows, we determine the
current version of krew from the installation receipt and clean all other
directories in $KREW_ROOT/store/krew that aren't the current version,
since we don't delete the install directory on post-upgrade if plugin
name=="krew" (due to inability of deleting currently executed binary on
win32).

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>

* add TestCleanupStaleKrewInstallations

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>

* fix lint error