pkg/index: ignore non-.yaml files during scanning
Turns out if there was a file in index/ directory with a non-.yaml extension,
we still targeted that as a plugin, so we tried to parse its manifest. (But
read/parsing of individual manifests do not cause fatal errors, just printed
to stderr like a warning.)

This creates a helper method that properly scans for .yaml files in an
immediate directory.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>