isSupportedAPIVersion's check is now relaxed
isSupportedAPIVersion function now checks to see if the input (apiVersion) exists within a list defined in pkg\constants\constants.go, if so it returns true, otherwise false is returned.

The function GetSupportedAPIVersions in constants.go is how the supported versions list is accessed. (It is hardcoded however, can be modified to use an external file instead)