Dynamic plugins list in the website (#658)
* Dynamic plugins list in the website

This relies on Netlify's serverless functions capability to render
plugin count (in the homepage) and plugin list (/docs/plugins) in
the website!

The serverless function is called "api" and it's accessible on path
/.netlify/functions/api/* and has two subpaths that return JSON responses:

1. */pluginCount
2. */plugins

This serverless function is built by placing a binary in /site/functions
named "api" and the build instruction is in 'netlify.toml' file.

Since we get plugin list by calling GitHub API, by default we run without
an authentication token. However after this is merged, I will go and add a
permissionless $GITHUB_ACCESS_TOKEN that I have created from my account.

Unauthenticated GitHub calls are limited to 60 per IP, but since I'll add a
personal access token, that should not be an issue. When debugging locally,
60 is also fine. (deploy previews for PRs from non-maintainers will not have
an access token available, so they will make unauthenticated calls to GH API.)

The responses from the dynamic functions are actually cached in Netlify's
CDN and that helps with the rate-limiting as well. I have currently set the
CDN cache duration to 1 hour for each response.

Please take a look at the previews and let me know if there is anything that
doesn't look good.

I might delete `generate-plugin-overview` tool and update the current
plugins.md in a followup PR.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>

* Apply review comments

* Use lit-html and async/await style for better readability

* Simplify work sharing with errgroup

* Satisfy impi constraints

* fix import

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>

* add setup/debugging documentation

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>

* move from /docs/plugins to /plugins

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>

* add /plugins to navbar

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>

Co-authored-by: Cornelius Weig <22861411+corneliusweig@users.noreply.github.com>