Changes install doc to remove question of authority
As the linux path prefers running what shows up first in the path, there is a potential for someone to create a script in .krew/bin to override another non-krew binary or script from anywhere in the system, even /bin. This puts krew at the end of the path to remove that potential and improve overall security.