Correct windows krew install command.
Running with PowerShell confuses a few users that we direct to install krew, as they come across this error
```
The command krew was not found, but does exist in the current location. Windows PowerShell does not load commands from the current location by default. If you trust this command, instead type: ".\krew". See "get-help about_Command_Precedence" for more details.
```
Suggest changing to `.\krew install krew` to avoid that